analitics remove from admin navber
