+++ frontend/src/components/Navbar.jsx
@@ -86,7 +86,7 @@ const Navbar = () => {
         <>
           <li><Link to="/admin/dashboard">Dashboard</Link></li>
           <li><Link to="/admin/reports">Reports</Link></li>
-          <li><Link to="/admin/analytics">Analytics</Link></li>
+          {/* <li><Link to="/admin/analytics">Analytics</Link></li> */}
           <li><Link to="/messages">Messages</Link></li>
           <li><Link to="/admin/management">Management</Link></li>
           {/* <li><Link to="/admin/validations">Management</Link></li> */}
