+++ frontend/src/pages/admin/Management.css
@@ -1,0 +1,324 @@
+/* Management.css */
+
+:root {
+    --primary: #2A3F5F;
+    --secondary: #4A90E2;
+    --accent: #FF5A5F;
+    --background: #F8FAFC;
+    --text: #2D3748;
+    --text-light: #718096;
+    --border: #E2E8F0;
+    --success: #48BB78;
+    --warning: #ECC94B;
+    --error: #F56565;
+    --shadow-sm: 0 1px 3px rgba(0,0,0,0.12);
+    --shadow-md: 0 4px 6px rgba(0,0,0,0.1);
+    --transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
+  }
+  
+  * {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    font-family: 'Segoe UI', system-ui, -apple-system, sans-serif;
+  }
+  
+  body {
+    background-color: var(--background);
+  }
+  
+  .management-container {
+    display: flex;
+    min-height: 100vh;
+    background-color: var(--background);
+  }
+  
+  /* Sidebar Styles */
+  .sidebar {
+    width: 280px;
+    background: var(--primary);
+    padding: 1.5rem;
+    position: fixed;
+    height: 100vh;
+    transition: var(--transition);
+    box-shadow: var(--shadow-md);
+  }
+  
+  .sidebar-menu {
+    list-style: none;
+    margin-top: 2rem;
+  }
+  
+  .sidebar-menu li {
+    color: white;
+    padding: 1rem;
+    margin: 0.5rem 0;
+    border-radius: 8px;
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    gap: 1rem;
+    transition: var(--transition);
+  }
+  
+  .sidebar-menu li:hover {
+    background: rgba(255,255,255,0.1);
+  }
+  
+  .sidebar-menu li.active {
+    background: var(--secondary);
+    box-shadow: var(--shadow-sm);
+  }
+  
+  .sidebar-menu li span {
+    font-weight: 500;
+    flex-grow: 1;
+  }
+  
+  .menu-arrow {
+    margin-left: auto;
+    opacity: 0.8;
+    font-size: 0.9rem;
+  }
+  
+  /* Main Content Styles */
+  .main-content {
+    flex: 1;
+    margin-left: 280px;
+    padding: 2rem;
+    max-width: 1200px;
+  }
+  
+  /* Validation Container */
+  .validation-container {
+    background: white;
+    border-radius: 12px;
+    padding: 2rem;
+    box-shadow: var(--shadow-md);
+  }
+  
+  .validation-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 2rem;
+  }
+  
+  .validation-header h3 {
+    color: var(--primary);
+    font-size: 1.5rem;
+  }
+  
+  .badge {
+    background: var(--accent);
+    color: white;
+    padding: 0.25rem 0.75rem;
+    border-radius: 20px;
+    font-size: 0.8rem;
+    font-weight: 600;
+  }
+  
+  .instructions-list {
+    list-style: none;
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+  }
+  
+  .instructions-list li {
+    display: flex;
+    gap: 1rem;
+    align-items: flex-start;
+    padding: 1.5rem;
+    border-radius: 8px;
+    background: rgba(74, 144, 226, 0.05);
+  }
+  
+  .instruction-icon {
+    background: var(--secondary);
+    color: white;
+    padding: 1rem;
+    border-radius: 8px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  .view-request-btn {
+    margin-top: 2rem;
+    background: var(--secondary);
+    color: white;
+    border: none;
+    padding: 1rem 2rem;
+    border-radius: 8px;
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+    font-weight: 600;
+    cursor: pointer;
+    transition: var(--transition);
+  }
+  
+  .view-request-btn:hover {
+    opacity: 0.9;
+    transform: translateY(-1px);
+  }
+  
+  /* User Management */
+  .user-management-container {
+    background: white;
+    padding: 2rem;
+    border-radius: 12px;
+    box-shadow: var(--shadow-md);
+  }
+  
+  .user-management-container h3 {
+    color: var(--primary);
+    margin-bottom: 1.5rem;
+  }
+  
+  .user-list {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+  }
+  
+  .user-item {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 1rem;
+    border-radius: 8px;
+    background: var(--background);
+    transition: var(--transition);
+  }
+  
+  .user-item:hover {
+    transform: translateX(4px);
+  }
+  
+  .delete-btn {
+    background: var(--error);
+    color: white;
+    border: none;
+    padding: 0.5rem 1rem;
+    border-radius: 6px;
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+    cursor: pointer;
+    transition: var(--transition);
+  }
+  
+  .delete-btn:hover {
+    opacity: 0.9;
+  }
+  
+  /* Criminal Database */
+  .criminal-database-container {
+    background: white;
+    padding: 2rem;
+    border-radius: 12px;
+    box-shadow: var(--shadow-md);
+  }
+  
+  .criminal-form {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    gap: 1rem;
+    margin: 2rem 0;
+  }
+  
+  .criminal-form input {
+    padding: 0.75rem;
+    border: 1px solid var(--border);
+    border-radius: 6px;
+    font-size: 1rem;
+  }
+  
+  .criminal-form button {
+    grid-column: span 3;
+    background: var(--success);
+    color: white;
+    border: none;
+    padding: 1rem;
+    border-radius: 6px;
+    cursor: pointer;
+    font-weight: 600;
+    transition: var(--transition);
+  }
+  
+  .criminal-form button:hover {
+    opacity: 0.9;
+  }
+  
+  .criminal-list {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+    gap: 1.5rem;
+  }
+  
+  .criminal-card {
+    background: var(--background);
+    padding: 1.5rem;
+    border-radius: 8px;
+    position: relative;
+    box-shadow: var(--shadow-sm);
+  }
+  
+  .criminal-card h4 {
+    color: var(--primary);
+    margin-bottom: 0.5rem;
+  }
+  
+  .criminal-card p {
+    color: var(--text-light);
+    font-size: 0.9rem;
+    margin: 0.25rem 0;
+  }
+  
+  .remove-btn {
+    background: var(--error);
+    color: white;
+    border: none;
+    padding: 0.5rem 1rem;
+    border-radius: 6px;
+    margin-top: 1rem;
+    cursor: pointer;
+    transition: var(--transition);
+  }
+  
+  .remove-btn:hover {
+    opacity: 0.9;
+  }
+  
+  /* Utility Classes */
+  .bold {
+    font-weight: 600;
+  }
+  
+  .mt-2 {
+    margin-top: 2rem;
+  }
+  
+  /* Responsive Design */
+  @media (max-width: 768px) {
+    .sidebar {
+      width: 100%;
+      height: auto;
+      position: relative;
+    }
+  
+    .main-content {
+      margin-left: 0;
+      padding: 1rem;
+    }
+  
+    .criminal-form {
+      grid-template-columns: 1fr;
+    }
+  
+    .criminal-form button {
+      grid-column: span 1;
+    }
+  }
